fix: correct integer division in physics expressions and String identity tests
Integer division
Five expressions used integer division where floating point was meant, so the
term they scaled silently vanished:

  CoefCalcCompute.java:2384   (1/16)  -> (1.0/16.0)
  CoefCalcCompute.java:2563   (1/8)   -> (1.0/8.0)
  CoefCalcCompute.java:3759   (2/3) and shells[i]/Z, both now floating point
  CoefCalcCompute.java:3877   (2/3)   -> (2.0/3.0)   in getWkMolecule
  MicroED.java:240            (4/3)   -> (4.0/3.0)   spherical crystal volume

getWkMolecule previously collapsed to a*binding*1000, which is why commit
8d37c7a -- widening sumZ from int to double to fix a truncation there -- could
not change any result. The plasmon term now contributes. MicroED's spherical
volume was three quarters of the correct value, and it clamps exposedVolume at
MicroED.java:1825.

The golden-file tests for a standard MX run are byte-identical before and after
these changes, which independently confirms this code is unreachable from an MX
experiment: the effect is confined to the electron/GOS paths.

String identity
BEAM_CIRCULAR was tested with ==, which worked only because the ANTLR parser
happens to store the interned literal. Any caller building the property map
programmatically got a rectangular beam, a different normFactor and different
doses throughout. Fixed in BeamGaussian, BeamTophat and BeamExperimentalpgm,
along with the same pattern in Version.REVISION and MicroED's two crystalTypeEM
comparisons.

Robustness
ExposureSummary.exposureComplete indexed imageDWD[-1] when no image completed,
reachable by an exposure whose angular resolution exceeds its wedge span. Now
guarded.

Examples
examples/SMXray2_example_input.txt read "AbsCoefCalc CIF", which can never
parse: Inputfile.g:393 defines the token named CIF as matching the literal
"EXPSM", while CIFNAME matches "CIF". Now "AbsCoefCalc EXPSM". It still cannot
be run -- it names an unshipped Fe3O4 CIF file and CoefCalcFromCIF reads local
paths only -- so it is excluded from the example sweep with a separate test
asserting its keyword is valid.

Tests
CrystalCuboidTest's 180-degree negation assertion was wrong about the rotation
convention and is replaced by a correct invariant: a full 360 degree rotation
leaves every voxel coordinate unchanged. Nothing is parked now; ant test-pending
reports zero tests. 246 tests pass.

+++ src/se/raddo/raddose3D/CoefCalcCompute.java
@@ -2381,7 +2381,7 @@ public class CoefCalcCompute extends CoefCalc {
     meanZoverA = 0.56;
     stoppingPower = K * (meanZoverA)* (1/betaSquared) * 
                     (Math.log((csquared*betaSquared*m*Math.pow(gamma-1, 0.5))/(meanJ * Math.pow(2, 0.5)))+ 
-                        0.5*(1-betaSquared)-((2*gamma-1)/(2*Math.pow(gamma, 2))) + (1/16)*Math.pow((gamma-1)/gamma, 2));
+                        0.5*(1-betaSquared)-((2*gamma-1)/(2*Math.pow(gamma, 2))) + (1.0/16.0)*Math.pow((gamma-1)/gamma, 2));
     stoppingPower = stoppingPower * 1000 * density /1E7;
     
     //without corrections
@@ -2560,7 +2560,7 @@ stoppingPower = stoppingPower * 1000 * density /1E7;
     double energy = avgEnergy * Beam.KEVTOJOULES;
     Fbeta = Math.log((m*csquared*(energy)* betaSquared) / (2*(1-betaSquared))) 
             - (2*Math.pow((1-betaSquared),0.5) - 1 + betaSquared)
-            * Math.log(2) + 1 - betaSquared + (1/8)*(1-Math.pow(1-betaSquared,0.5));
+            * Math.log(2) + 1 - betaSquared + (1.0/8.0)*(1-Math.pow(1-betaSquared,0.5));
     stoppingPower = (0.153536/betaSquared)*(sumZ/sumA)*(Fbeta - 2*Math.log(meanJ) - delta);
     stoppingPower = stoppingPower * 1000 * passedDensity /1E7;
     double radiativeStopping = ((KE_MeV*meanZ)/800)*stoppingPower;
@@ -3756,7 +3756,7 @@ stoppingPower = stoppingPower * 1000 * density /1E7;
       //get Wi
       if (numInnerShells > 0) {
         for (int i = 0; i < numInnerShells; i++) { 
-          Wi[i] = Math.pow(Math.pow(a * getShellBinding(i, e)*1000,2) + (2/3)*(shells[i]/Z)*Math.pow(plasmaEnergy, 2), 0.5);
+          Wi[i] = Math.pow(Math.pow(a * getShellBinding(i, e)*1000,2) + (2.0/3.0)*((double) shells[i]/Z)*Math.pow(plasmaEnergy, 2), 0.5);
         }
       }
       
@@ -3874,7 +3874,7 @@ stoppingPower = stoppingPower * 1000 * density /1E7;
     }
     double plasmaEnergy = getPlasmaEnergyAll(surrounding);
   //  double Wk = Math.pow(Math.pow(a * getShellBinding(shellIndex, e)*1000,2) + (2/3)*((fk*totNum)/(sumZ))*Math.pow(plasmaEnergy, 2), 0.5);
-    double Wk = Math.pow(Math.pow(a * getShellBindingSubshell(shellIndex, e)*1000,2) + (2/3)*((fk*totNum)/(sumZ))*Math.pow(plasmaEnergy, 2), 0.5);
+    double Wk = Math.pow(Math.pow(a * getShellBindingSubshell(shellIndex, e)*1000,2) + (2.0/3.0)*((fk*totNum)/(sumZ))*Math.pow(plasmaEnergy, 2), 0.5);
     //I am really guessing the Lorenz Lorentz sumFk sum Z but just a hunch
     return Wk;
   }
